hack in audio file support
note: this is actually bad. The way I hacked it in is to generate a
thumbnail based on the name in the thumber. But, that means it's not
actually trying to load the audio file to see if it can actually load
it.

further, on viewer I generate a new image every time the file is
displayed and hacked the code to show both a picture, and the video
controls.

Instead I should probably show HTML if audio.